trigger::action module broken up into more manageable submodules and a separate tests file.
